pip_requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.0
argcomplete==3.6.2
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.13.3
bitarray==3.3.1
bitsandbytes==0.45.5
bleach==6.1.0
braceexpand==0.1.7
Brotli==1.0.9
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
ci_sdr==0.0.2
click==8.1.8
cmake==4.0.0
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.2
ctc_segmentation==1.7.4
cycler==0.12.1
Cython==3.0.12
datasets==3.1.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
editdistance==0.8.1
einops==0.8.1
email_validator==2.2.0
encodec==0.1.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fairseq==0.12.2
fast_bss_eval==0.1.3
fastapi==0.115.4
fastapi-cli==0.0.5
fastdtw==0.3.4
ffmpy==0.3.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.2.3
funasr==1.0.27
g2p-en==2.1.0
Gammatone==1.0
gdown==4.4.0
gitdb==4.0.12
GitPython==3.1.44
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface_hub==0.26.5
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.0.1
hypy_utils==1.0.29
idna==3.10
importlib-metadata==4.13.0
importlib_resources==6.4.5
inflect==7.4.0
ipywebrtc==0.6.0
jaconv==0.4.0
jamo==0.4.1
jellyfish==1.1.3
jieba==0.42.1
Jinja2==3.1.4
jiter==0.9.0
jiwer==3.1.0
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaggle==1.7.4.2
kaldiio==2.18.1
kiwisolver==1.4.2
laion_clap==1.1.6
lazy_loader==0.4
Levenshtein==0.25.1
librosa==0.10.1
lightning==2.3.3
lightning-utilities==0.11.9
llvmlite==0.41.1
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
mdurl==0.1.2
mir_eval==0.8.2
mock==5.2.0
modelscope==1.25.0
more-itertools==10.5.0
mpmath==1.3.0
msclap==1.3.3
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
narwhals==1.35.0
networkx==3.1
nltk==3.9.1
nnAudio==0.3.3
nose==1.3.7
numba==0.58.1
numpy==1.22.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.68.2
openai-whisper==20240930
opt_einsum==3.4.0
orjson==3.10.7
oss2==2.19.1
packaging==24.2
pandas==2.0.3
pesq==0.0.5
pillow==10.2.0
pip==25.0.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.0.0
progressbar==2.5
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.44.0
pypinyin-dict==0.9.0
pysepm==0.1
PySocks==1.7.1
pysptk==1.0.1
pystoi==0.4.1
python-dateutil==2.9.0
python-multipart==0.0.18
python-slugify==8.0.4
pytorch-lightning==2.4.0
pytorch-wpe==0.0.1
pytz==2024.2
pyworld==0.3.5
PyYAML==6.0.2
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rotary-embedding-torch==0.8.6
rpds-py==0.20.1
ruff==0.3.5
s3prl==0.4.17
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.3.7
speechmos==0.0.1.1
SRMRpy==1.0
starlette==0.41.2
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.4.1+cu121
torch-complex==0.4.4
torchaudio==2.4.1+cu121
torchlibrosa==0.1.0
torchmetrics==1.5.2
torchvision==0.19.1+cu121
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typeguard==4.4.0
typer==0.12.5
typer-slim==0.12.5
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.7
unicodedata2==15.1.0
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.32.1
versa-speech-audio-toolkit==0.0.1
wandb==0.19.9
wcwidth==0.2.13
webdataset==0.2.100
webencodings==0.5.1
websockets==11.0.3
wget==3.2
wheel==0.44.0
xmltodict==0.14.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.15.2
yq==3.4.3
zipp==3.20.2
zstandard==0.23.0
